server/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
anthropic==0.18.1
google-generativeai==0.3.2